# src/api.yaml
openapi: 3.0.0
info:
  title: Employee management system API
  description: Api Testing.
  version: 1.0 
  

# servers:
#   - url: http://localhost:8000/
#     description:  Local server 
  # - url: https://prod.com/
  #   description:  Pre Production server
  # - url: https://test.com/
  #   description:  Production server
components:
  securitySchemes:
    token:        # arbitrary name for the security scheme
          
      type: http
      scheme: bearer
    
    # ApiKey:        # arbitrary name for the security scheme
    #   type: apiKey
    #   in: header       # can be "header", "query" or "cookie"
    #   name: apikey

# ----------------------------------- Authentication swagger api --------------------------- #

#login api
paths:
...
  /api/user:
    post:
      tags:
        - user
      security:
       - token: []
      summary: Returns a success message.        
      post:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                employee_id:         
                  type: String
                  example : "D9-1"
                first_name:       
                  type: string
                  example : "abc"
                last_name:       
                  type: string
                  example : "xyz"
                email:       
                  type: string
                  example : "xyz@gmail.com"
                password:       
                  type: string
                  example : ""
                phone:
                  type: integer
                  example : "1234567890"
                report_by:       
                  type: string
                  example : ""
                role_id:       
                  type: string
                  example : ""
                department_id:       
                  type: string
                  example : ""
                designation_id:   
                  type: string
                  example : ""
                joining_date:       
                    type: string
                    format: date
                    example : "2023-09-01"
                status:
                    type: "string"
                    enum: ['Active', 'Inactive']
              required:
                - employee_id
                - first_name
                - last_name
                - email
                - phone
                - report_by
                - joining_date
                - status
                - password
                - role_id
                - department_id
                - designation_id
         
      responses:
        '201':
          description: Success message. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

# all user get
...
  /api/user/:
    get:
      tags: ["user"]
      security:
       - token: []
      summary: Returns a user details.
      responses: 
        '200':
          description: Success message
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.

# get single user
...
  /api/user/{id}:
    get:
      tags: ["user"]
      security:
       - token: []
      summary: Returns a user details by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
      responses: 
        '200':
          description: Success message
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.
 #  update user
    put:
      tags: ["user"]
      security:
       - token: []
      summary: user detail update.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
      put:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                address:         
                  type: String
                  example : ""

      responses: 
        '200':
          description: Success message
        '400':
          description: Bad Request
        '404':
          description: Not found
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.

 #  delete user
    delete:
      tags: ["user"]
      security:
       - token: []
      summary: user detail delete.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
  
      responses: 
        '200':
          description: Success message
        '400':
          description: Bad Request
        '404':
          description: Not found
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.
 #  update status user
    patch:
      tags: ["user"]
      security:
       - token: []
      summary: user status update.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
  
      responses: 
        '200':
          description: Success message
        '404':
          description: Not found
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.

#  check email
...
  /api/user_document:
    post:
      tags:
        - User document
      security:
       - token: []
      summary:  upload document for user.     
      post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resume:
                  type: string
                  format: binary
                offer_letter:
                  type: string
                  format: binary
                joining_letter:
                  type: string
                  format: binary
                other:
                  type: string
                  format: binary
                user_id:
                  type: string
         
      responses:
        '200':
          description: Success message. 
        '201':
          description: created successfully. 
        '400':
          description: Bad Request.
        '404':
          description: Not Found.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

##--------------------------------------------  emergency contact api -----------------------------------------##
...
  /api/document:
    post:
      tags:
        - document
      security:
       - token: []
      summary:  add document.     
      post:
      requestBody:
        required : true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                name:
                  type: string
                  example : ""
                description:
                  type: string
                  example : ""
         
      responses:
        '201':
          description: created successfully. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

# get holiday
    get:
        tags:
         -  document
        security:
        - token: []
        summary: list document.   
          
        responses:
          '200':
            description: Success message. 
          '401':
            description:  Unauthorized.
          '403':
            description: Access Denied.
          '500':
            description: Server error.

# update  api
...
  /api/document/{id}:
      put:
        tags:
          -  document
        security:
        - token: []
        summary: update document.
        parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.        
        put:
        requestBody:
          required : true
          content:
            multipart/form-data:
                  schema:
                    type: object
                    properties:
                      image:
                        type: string
                        format: binary
                      name:
                        type: string
                        example : ""
                      description:
                        type: string
                        example : ""
          
        responses:
          '200':
            description: ok. 
          '400':
            description: Bad Request.
          '404':
            description: Not Found.
          '403':
            description: Access Denied.
          '401':
            description:  Unauthorized.
          '500':
            description: Server error.

            # delete docuement
      delete:
        tags:
          -  document
        security:
        - token: []
        summary: delete document.
        parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.     
          
        responses:
          '200':
            description: ok. 
          '400':
            description: Bad Request.
          '404':
            description: Not Found.
          '403':
            description: Access Denied.
          '401':
            description:  Unauthorized.
          '500':
            description: Server error.


#  ------------------------------------------------ Dashboard api ----------------------------------------------#
...
  /api/role:
    post:
      tags:
        - user role
      security:
       - token: []
      summary: add user role.        
      post:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                name:       
                  type: string
                  example : ""
                permissions:       
                  type: array
                  example : [{
                                  "menuId": "64f5c3fac742a9d98546e2e9",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                              },
                              {
                                  "menuId": "64f5c4cbc742a9d98546e2ed",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                              },
                              {
                                  "menuId": "64f5c4dac742a9d98546e2f1",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                              },
                              {
                                  "menuId": "64f5c4e2c742a9d98546e2f5",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                              },
                              {
                                  "menuId": "64f5c4f6c742a9d98546e2f9",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                              },
                              {
                                  "menuId": "64f5c4fbc742a9d98546e2fd",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                              },
                              {
                                  "menuId": "64f5c500c742a9d98546e301",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                              },
                              {
                                  "menuId": "64f6e7d9de634a7f7c48a2cb",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                              },
                              {
                                  "menuId": "6503e4540f87047dfd98ecb0",
                                  "list": 0,
                                  "create": 0,
                                  "delete": 0,
                                  "update": 0,
                          }]
              required:
                - name
                - permissions
         
      responses: 
        '201':
          description: created sucess. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '500':
          description: Server error.

# get user role
    get:
        tags:
         -  user role
        security:
        - token: []
        summary: list user role.   
          
        responses:
          '200':
            description: Success message. 
          '401':
            description:  Unauthorized.
          '403':
            description: Access Denied.
          '500':
            description: Server error.

# update  api
...
  /api/role/{id}:
      put:
        tags:
          -  user role
        security:
        - token: []
        summary: update user role.
        parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.        
        put:
        requestBody:
          required: true
          content:
          # multipart/form-data:
            application/json:
              schema:
                properties: 
                  name:       
                    type: string
                    example : ""
                  permissions:       
                    type: array
                    example : [{
                                    "menuId": "64f5c3fac742a9d98546e2e9",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                                },
                                {
                                    "menuId": "64f5c4cbc742a9d98546e2ed",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                                },
                                {
                                    "menuId": "64f5c4dac742a9d98546e2f1",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                                },
                                {
                                    "menuId": "64f5c4e2c742a9d98546e2f5",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                                },
                                {
                                    "menuId": "64f5c4f6c742a9d98546e2f9",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                                },
                                {
                                    "menuId": "64f5c4fbc742a9d98546e2fd",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                                },
                                {
                                    "menuId": "64f5c500c742a9d98546e301",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                                },
                                {
                                    "menuId": "64f6e7d9de634a7f7c48a2cb",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                                },
                                {
                                    "menuId": "6503e4540f87047dfd98ecb0",
                                    "list": 0,
                                    "create": 0,
                                    "delete": 0,
                                    "update": 0,
                            }]
                required:
                  - name
                  - permissions
          
        responses:
          '200':
            description: ok. 
          '400':
            description: Bad Request.
          '404':
            description: Not Found.
          '403':
            description: Access Denied.
          '401':
            description:  Unauthorized.
          '500':
            description: Server error.

# get 
      get:
          tags:
          -  user role
          security:
          - token: []
          summary: user role get by id.  
          parameters:
            - name: id
              in: path
              required: true
              example : "static"
              description: Parameter description in CommonMark or HTML.       
              
          responses:
            '200':
              description: Success message. 
            '401':
              description:  Unauthorized.
            '403':
              description: Access Denied.
            '500':
              description: Server error.

# check 
...
  /api/role/name:
      post:
        tags:
          -  user role
        security:
        - token: []
        summary: check user role name.        
        post:
        requestBody:
          required: true
          content:
            # multipart/form-data:
            application/json:
              schema:
                properties: 
                  name:       
                    type: string
                    example : ""
                required:
                  - name
          
        responses:
          '200':
            description: ok. 
          '400':
            description: Bad Request.
          '401':
            description:  Unauthorized.
          '500':
            description: Server error.
          '403':
            description: Access Denied.


#  -------------------------------------------  project api -------------------------------------- #
# get 
...